refactor: unused function
