+++ packages/shared/src/components/squads/SharePostBar.tsx
@@ -15,6 +15,7 @@ import { useLazyModal } from '../../hooks/useLazyModal';
 import { LazyModal } from '../modals/common/types';
 import { Squad } from '../../graphql/sources';
 import { ExternalLinkPreview } from '../../graphql/posts';
+import { TutorialKey, useTutorial } from '../../hooks/useTutorial';
 
 export type NewSquadPostProps = Pick<
   PostToSquadModalProps,
@@ -37,6 +38,18 @@ function SharePostBar({
   const [url, setUrl] = useState<string>(undefined);
   const isMobile = !useMedia([mobileL.replace('@media ', '')], [true], false);
   const onSharedSuccessfully = () => setUrl(undefined);
+  const sharePostTutorial = useTutorial({
+    key: TutorialKey.ShareSquadPost,
+  });
+
+  const copyLinkTutorial = useTutorial({
+    key: TutorialKey.CopySquadLink,
+  });
+  const onCompleteTutorials = () => {
+    sharePostTutorial.complete();
+    copyLinkTutorial.activate();
+  };
+
   const onOpenCreatePost = (preview: ExternalLinkPreview, link?: string) =>
     openModal({
       type: LazyModal.CreateSharedPost,
@@ -44,6 +57,7 @@ function SharePostBar({
         preview: { ...preview, url: link },
         squad,
         onSharedSuccessfully,
+        onAfterClose: onCompleteTutorials,
       },
     });
 
@@ -53,6 +67,7 @@ function SharePostBar({
       props: {
         onArticleSelected: ({ post }) => onOpenCreatePost(post, post.permalink),
         keepOpenAfterSelecting: true,
+        onAfterClose: onCompleteTutorials,
       },
     });
